Log idea.log from debug function
debugEmitLogs has no callers and is only intended to be used when
someone needs to debug an issue locally. In that scenario, it's common
to want idea.log to be outputted, so now it's done by default without
requiring a _second_ code change.

This change is essentially just to save people from having to read even
more documentation.

Bug: N/A
Test: N/A - this method is unused and must be invoked manually
Change-Id: Ia5492fa653f29c67c4ef8d991a25ffb9a6d9198b

GitOrigin-RevId: fa016ee7dfbbbe650598c83597b8586737f545df