Ensure root libraries have runtime dependencies
When the option to derive runtime classpaths for library modules is
enabled we need to ensure that libraries that aren't used by any app
have there runtime classpath requested explicitly. This is because we
will not be able to derive this information from the transitive
dependency graphs of other projects.

To do this we ensure that we track whether or not projects that we
request are the root projects and if they are we always request the
runtime classpath.

Bug: 267066187
Test: snapshots updated showing the new runtime classpath
Change-Id: I67b985a90642a82dc2969b0170c0b2667e54b1f5

GitOrigin-RevId: 8241952359f6fb3599f40fdca05ab56867fc448e