Ensure ComboBoxes do not pop-in UI
This change pre-sets the text for two ComboBoxActions so that
we don't need to wait for a call to update() to get the text.
If we ever decide to change the translations editor to bring back
the state from when it was closed, we'll need to update this as well.

Fixes: 288284518
Test: Filter{Keys,Locales}ActionTest.kt
Change-Id: Ibbd60481e35521675c00ad77ee807d178867837d

GitOrigin-RevId: 75cb4b1654c2d2059952a706b01b861130f97619