[BA] Make WindowsDefenderChecker copy secure
Following the discussion with the author of the original
WindowsDefenderChecker I see the risks of opening this API broader
than necessary. Updating this copy to match the visibility of the
original, so that only we can use it.

Also remove unnecessary Extension class, the one from the original
is used.

Bug: 285867050
Test: Existing
Change-Id: Icd98ee1d1f61fafb5a9febe575d96bbb2dec2cb2

GitOrigin-RevId: a05e62887c5e764773bfcd7fcb7a4364c54a18bf