Update background for IconLabels
Previously, the updateTablePresentation method didn't update the
foreground and background of icons, since they didn't matter; the
background is only drawn if the component is opaque, and the foreground
is not used. However, with the addition of hovering support, the
component under the mouse may be opaque, which results in its background
being drawn on top of the row's background, so it needs to be the
correct color.

Since IconTableComponent is an interface that can't enforce that it's in
a JComponent, push the implementation of updateTablePresentation down to
the concrete classes.

Bug: 287348134
Test: updated
Change-Id: I0cf9d728753a3c5c3362502f3df6e1432a4b20c6

GitOrigin-RevId: fde00e4cb2b993df6746abf2dffa09d43bdffcb3