Remove RequestModelUpdate from LayoutlibSceneManager
The only place where this was used was expecting a re-render to
happen after the model update. However, a previous change that
removed some unnecessary re-renders, it also removed this needed
re-render by accident.

To avoid problems in the future, instead of putting the re-render
logic back in the model change listeners, this change makes a
re-render to explicitly happen in activation.

Fixes: 263849840
Test: manually verified
Change-Id: I7c3580d9f20e6a1ba9234bf95b07b25a6fe1e512

GitOrigin-RevId: ece9da5896cf6a995a30fa91a7e382decf25f3bc