Do a new render after running callbacks in Compose Preview
Some Composables do not render properly after their first render as they
rely on callbacks to layout and measure their final state. For that
reason, we need to run callbacks after the first render, and request a
new render so that the results of running the callbacks are displayed in
the Preview.

Fixes: 287312025
Fixes: 265668716
Test: checked that it renders correctly the examples from the bugs above
Change-Id: If94e0b1939fcaab80cabf30b164b153b953548ea

GitOrigin-RevId: 9ba0f634913ecfe0b2a5bd445f8fa6866a062f78