IDEA-323476 Update default value for BUILD_OUTPUT_DOWNLOADS_INFORMATION
Modified the accessor for the `BUILD_OUTPUT_DOWNLOADS_INFORMATION` StudioFlag from a simple property get to a method call `isBuildOutputShowsDownloadInfo()`. In this case, it checks if the Flag is overridden, if not it defaults to `IdeInfo.getInstance().isAndroidStudio()`. This ensures that the feature is not enabled by default in IDEA while maintaining a possibility to be changed by a user.
 Move flag check to an outer level, to prevent unnecessary code execution when the flag is disabled.

Merge-request: IJ-MR-109445
Merged-by: Dmitry Yudin <Dmitry.Yudin@jetbrains.com>
(cherry picked from commit aff1eb9e77c9bedce7fde14c70980db3b13072c4)

IJ-MR-109445

GitOrigin-RevId: 7edac5d8c608279e0274f90f06d8c8f8bcf08b01