Don't pass null to AndroidBundle.message
The NotNull annotation on the varargs appears to be treated differently
in monobuild vs the normal studio build. In the monobuild it is treated
as a element wise non nullable check which fails given the arguments in
AndroidGradleProjectSettingsControlBuilder.kt. Conversely in the normal
studio build it appears to be treated as the vararg array not being
null.

This change ensures that the arguments are not null to satify both
cases.

Bug: N/A
Test: N/A
Change-Id: Id4dcf98581e44be199b4fae147e8490ba89ef569

GitOrigin-RevId: dcf650311cbb44576e47b975917860edd6ec9dad