[IDEA-324336] Refactor AndroidRunConfigurations service to project service and resolve app freezing issue.
When the LaunchUtils#isWatchFeatureRequired method is called from the Event Dispatch Thread (EDT), an error is thrown due to accessing VFS on the EDT.
This issue has been fixed by running VFS operations outside of the EDT thread.

In the production environment, when setting up an Android run configuration, the AndroidRunConfigurations#setupRunConfigurationsBlocking method now launches
a new coroutine. The required data is computed outside of the EDT thread within a read action.
Subsequently, the run configurations are created within a write action on the EDT thread.

For headless and test environments, the AndroidRunConfigurations#setupRunConfigurationsBlocking method follows a similar approach.
The required data is computed under a read action, and the configurations are created under a write action.
Both actions are executed within the same thread as the setupRunConfigurationsBlocking method to prevent any potential deadlocks.

GitOrigin-RevId: 0df93cab1deaff35f5883a05e83010ffa9b70ded