Remove AndroidGradleProjectDumbStartupActivity
All this does is disable three notifications emitted by the platform.
None of which really need to be disabled anymore.

Two of them are emitted on Jvm setup which we skip and the last one are
configuration failures from old versions of Gradle encountered during
running tests, these we likely want to display.

This way of siliencing notificationis is also very problematic since it
will disable them 'forever' until the user re-enables them in the
settings.

Bug: N/A
Test: N/A
Change-Id: I1c9399a923414e11b1f584b2d17511836d84a3e5

GitOrigin-RevId: 39ce0db7af7605f8806e0b0703adc4f8a30338bf