Listen to PROJECT_SYSTEM_SYNC_TOPIC
AndroidProjectRootListener needs to notify consumers when module roots
change, or more specifically when the resource folders associated with a
module may change. The current event in use does not fire at all times
when sync completes, so this change is adding in a second event that
should help capture more cases.

The consumers that are invalidated try to be smart about only
invalidating folders that have actually changed, so I don't anticipate a
larger perf impact to invalidating more frequently than we used to.

Fixes: 286877308
Test: AndroidProjectRootListenerTest
Change-Id: I357f9f25c733b74add2829c870437763a23987d9

GitOrigin-RevId: ad73fa8f2f31a01ca3b40af50597edf5778167ac