Replace implementation of ViewNode.flatten
Local tests shows that replacing the flatten sequence with a list
implementation makes ViewAndroidWindow.createDrawChildren() about
10 times faster.

The bug 290782309 is less likely to happen with this change.
At worst the wait freeze time would be 2.4 seconds rather than 24
seconds. At best this also reduces the chance of GC in the middle of
this function.

Fixes: 290782309
Test: Existing tests
Change-Id: I247f398c2b0efc0f24940b1908d1ef3009be08f0

GitOrigin-RevId: 84759fa2c0b73c68f64d84e6f52b48d22d4e0daa