Fix breakpoint setting for composable singletons.
When attaching to a running process where all the classes have
already been loaded, we only considered a subset of the subclasses
of the composable singletons.

This is because `vm.nestedTypes` only gets you one level
of nested classes. The fix is to recursively compute all the
nested classes so that we find the class that contains the
location on which to set the breakpoint.

This all works when starting the app from android studio
because that hits the path that ask the VM to notify us as
classes are being loaded. For that, we use a regular expression
that matches all of the subclasses.

Bug: 289283336
Test: Manual

Change-Id: Ied768b85392f5bcdd0ebf68b3bfda11defa5fc98

GitOrigin-RevId: 515a8bc888770640b401855832377499a515e104