Do not throw in getModuleIdForSyncRequest
I've hit this exception a few times when opening a project. I can't
reproduce reliably though. From the stack trace it seems to happen when
sync is somehow triggered by the platform, and is due to modules not
being Gradle modules.

The call to getGradleProjectPath has 7 branches where null is returned.
There are likely edge cases where one of these branches will trigger.
Since this code is used pre-sync to collect which variants we request I
think it makes sense for this not to throw. Throwing here terminates
sync and leaves the user in a very bad state. Project import can be
completed correctly even if this returns null.

Bug: 291323643
Test: N/A
Change-Id: I1d56a76b0b8384b5eb224d7a9d8e155be7ae0fd7

GitOrigin-RevId: a7e1e10d6a95b28a12d6a2bc70638a5ba1066915