Fix ClassLoadingIssueChecker class not found error reporting
Similarly to previous change it seem to have been broken during
previous migration. Based on the code existed before, class not found
pattern is not necessary but optional. Normally there is just a class
name in the exception message from what I have seen in my tests, but I
assume there might environment be versions when 'not found.' suffix
is added.

Additionally, fix build output consume method, refactor some lines and
convert tests for this checker to be integration tests.

Bug: 292231180
Test: Migrated ClassLoadingIssueCheckerTest as integration test
Change-Id: I680993e57f0a6b7e6a6c0d5bd73402afd60c7088

GitOrigin-RevId: 30e2508564fca5fcc7457a113991581c3c7bd158