Remove manual scrolling logic in stack trace
I don't know why this was here in the first place, but I think it was to
work around some scrolling issue. However that issue isn't present
anymore.

This logic was causing the editor's scroll pane to jump to the bottom
because the caret was moved to the bottom upon mouse wheel scroll for
reasons unknown (and it only happens on some stack traces).

Since the editor is read only, I took the liberty to disable the caret
in this change. It still jumps to the bottom for certain stack traces
but it isn't visible and doesn't cause scroll. Users' ability to select
text is not affected.

Test: manually - tried writing a test with fake ui but failed.
Fixes: 272733801
Change-Id: If63ceec7ce82ff1d2109bcdcd92f973a19966366

GitOrigin-RevId: 273966e722a8f81789e1fff7353e9a6033c16a4d