Show error message tips in a bottom panel
Tip messages in RenderErrorContributor are messages that
shows a possible solution for a given error.
The icons for those tips were wrapped into an <img> html
a the very end of the error stacktrace.
Wrapping images into html is not working for svgs.

This change extracts all the tip messages into a separate
panel in the bottom of the error starcktrace.
It also extracts the images from the html wrapper
and it shows it into a JPanel instead.

Bug: 267672487
Test: RenderErrorContributorTest, ComposeRenderErrorContributorTest
Change-Id: I32b2618dcfe318ced8379d971d3ffa442116924b

GitOrigin-RevId: 6ad89b089e1378428af03a9acd0a00beceb0f646