Remove spurious tooltip in component tree
The implementation of TreeTable is explicitly setting the table
tooltip to whatever the last generated renderers tooltip is.
This is usually the last item in the component tree.

Override the table tooltip such that we never use this value, but
instead have a custom implementation of getting a tooltip by mouse
location.

Fixes: 287929757
Test: Added unit test
Change-Id: I1cc774d8cf63b498c4c4a5fc5be849a83427928d

GitOrigin-RevId: 294cbc2a2112ef57ca9051532e9701d0c44c99fc