Make sure that DEVICE_HANDLE_KEY is registered
If AnAction.update() uses a DataKey to look up an object from its
DataContext, that DataKey needs to be registered before update() is
called, or else the data may not be present in the DataContext (at least
in the context of an ActionGroupPopup, which apparently creates a
PreCachedDataContext).

Move the DEVICE_HANDLE_KEY to a class that is going to be
loaded before anything in Studio get a DeviceHandle, and also reference
the keys in DeviceRowData's constructor so that they are loaded in tests
(which may not use DeviceProvisionerFactory).

Note that loading of top-level fields is rather subtle: merely
instantiating a class in the same file is insufficient to cause them to
be loaded. They are only loaded when a top-level field or method in the
file is referenced, causing the container class (e.g.
DeviceProvisionerFactoryKt to be loaded; moving the declaration to
DeviceProvisionerFactory only has an effect because there is another
top-level field that is referenced when PhysicalDeviceProvisionerFactory
is created. Previously, DEVICE_ROW_DATA_KEY was loaded only because of
the titlecase() method being called, not just because a DeviceRowData
was created.

In short, be very careful about top-level declarations that have
side-effects.

Fixes: 290067332
Test: Added
Change-Id: I5a09c0d7db593bed04b4d1b16c91cf3e99280b27

GitOrigin-RevId: a3b5dba2852cdd0d079d89fbcbb16d75385afd2b