Combine ResourceFolderManager main and test folder
In order to better support resources in androidTest modules, we need to
be able to identify what folders those resources live in. With linked
modules, there's no longer a need for ResourceFolderManager to
distinguish between app and test resources for a single module, since
each will be exposed by its relevant module/facet. By exposing the
androidTest resource folders as the "main" folders for its module, we
get a lot of the resource infrastructure closer to working with test
resources.

The substantive change is in `ResourceFolderManager.readFromFacetState`,
and in the removal of the `testFolders` property. All the rest of this
change is effectively updating to reflect that.

In terms of usage, the effect of the change is that whenever consumers
call `ResourceFolderManager.folders` for an androidTest module, they
will now get that module's folders instead of its linked main module's
folders.

Downstream there are slight changes included to
`ModuleResourceRepository.forTestResources` as well. This method is only
used for R class generation, and all the relevant R classes are still
created via each respective module, so there's no drop in functionality.

This change addresses b/265199079: go to definition from a Kotlin file
in androidTest to the corresponding resource definition in androidTest
will now work correctly.

Fixes: 265199079
Test: TestRClassesTest
Change-Id: I46d92f9391528ff00c7f546e407b9c681b7d9648

GitOrigin-RevId: 0e8d1b74316305cc5906ee4c0506aff5b6a83942