Use ClassId in {find,has}Annotation.
These utility functions in AndroidKtPsiUtils call into the Analysis API
under K2. The versions of these functions that accept strings or FqNames
cannot properly handle nested annotation classes in K2 mode, because K2
is expecting a ClassId. To avoid these issues, create a version of these
functions that accepts a ClassId, and deprecate the string and FqName
versions in favor of the ClassId version.

Bug: 278637936, 279139016
Test: Treehugger

Change-Id: I336f6c31bfa427c09d6f017a1a738c90d1a614d5

GitOrigin-RevId: 328efb9318abcd3cad18ae8174c6a8314833e685