Fix Gallery tabs disappearing when adding/removing @Preview
Adding and removing @Preview annotations from code triggers a refresh
that will eventually add/remove a corresponding tab to Gallery tabs.
The operation of adding (i.e. replacing) the updated toolbar was not
being guarded by a write intent lock, so ActionUpdater would throw an
exception and the tabs would not being added.

This change fixes the issue by wrapping the addition of the toolbar
into an invokeLater block, ensuring it will be executed under the
write intent lock.

Fixes: 291935659
Test: GalleryTabsTest, ComposeGalleryModeTest. Also manully verified.
Change-Id: I664d5eeae47b8fa53afca76eba37d0e4444e3ec1

GitOrigin-RevId: 9d6be87560f048fee0c88bd457b4a7be9eda2dd6