Keep Preview Gallery selection if tab is only updated
If we're editing a @Preview (e.g. adding a new parameter), we should
keep the gallery selection, even though the underlying object that
represents the tab will change. This provides a better experience to
the user, as they will keep previewing the same preview as before
instead of having the first tab selected by default.

For exemple, if a user has a tab selected and change the value of the
showSystemUi parameter, we can modify the preview and they'll see this
change instantly, instead of having to reselect the tab.

Fixes: 291938388
Test: GalleryTabsTest
Change-Id: I8a8e6e46eec40cb286c1de54c9c1d55cfba3a04d

GitOrigin-RevId: 1d6c2e7f7dd791fcb357ee29e1a596012f86c2b4