Enable/Disable Facet Tool Windows Dynamically
The Tool Windows in this cl used `facet.toolWindow` to
enable/disable themselves. This does not work well in IntelliJ because
the facet is tested only once when the project is loaded. But in
IntelliJ, a non Android project can become an Android project by adding
an Android module. With the current code, this project will not have
any Android Tool Windows until it is reloaded.

The ``<library.toolWindow>` extension point solves this by reevaluating
the applicability of a Tool Window whenever the project structure
changes.

Two new 'LibrarySearchHelper` are added in this cl to match the facets
that each tool window has declared.

Bug: 277937502
Test: NA
Change-Id: If830f3f278a5b215374d973da9316a139b8d4bd1

GitOrigin-RevId: 06251b74b55b7f2ee4dbc398963af49b60e0da52