Refactor issue checkers for AGP versions errors
Instances of AndroidSyncException identify failures in model builders.
They are not thrown from the build as usual exceptions but caught in
model builders, passed back as a model and recreated and rethrown on
studio side.
One set of these exceptions are about detected AGP version problems,
this change is about cleaning up the issue checkers for these.
Did some renaming and comments to make is easier understand what
the code does.
Migrated test to be integration test to make sure the whole process
handles it correctly.
In case of this test I use SimulatedSyncErrors which is not ideal
but such simulated exceptions are thrown in a similar way on studio
side as recreated AndroidSyncExceptions. I also make sure that they
follow the same procedure of converting to model and back as if it would
be passed from the model builders. So the difference is that in real
world exception would be thrown post-sync (simulated exceptions
are thrown before sync) and that it does not test actual checks of
the versions that generate exceptions.

Bug: 292231180
Test: Created AgpVersionExceptionsTest as integration test
Change-Id: I1bd07bd93e8ac5290dd96fc30271e001a7b02d6d

GitOrigin-RevId: 57bcdb417276693e5c373cebf5078ee203bb81d9