Add a general settings page for Essentials Mode
Requires some small tweaks to the Essentials Notifier to ensure that the
Essentials Mode notification expires if the mode is switched in the
general settings page. When providing the project on initial creation of
the notification, the notification does not expire when not passing in
the project on the expire call.

Fixes: 291141921
Tests: manual
Change-Id: I2d8e4a3a5d438176ac2aa19bbfbebe919f6c2fa7

GitOrigin-RevId: a734562c114d6a828b99a0b43c0f73e475c0d34c