Expose DeviceId from DataProvider
DeviceId is available as a public property in RunningDevicePanel.
This change also makes it available through RunningDevicePanel's
DataProvider.

This is useful for clients of RDW. It will allow them to get the
DeviceId directly from AnActionEvent and from the Contents,
without having to cast them to RunningDevicePanel.

For clients of RDW this is also advantageous in tests, where a
DataProvider will be sufficient, instead of implementing
RunningDevicesPanel. Which currently is not possible because it
contains internal properties.

Test: NA
Bug: 291971611
Change-Id: I78af32012730179231119a02e4a25549004a3f90

GitOrigin-RevId: bd6b7fc639d01732b120bfda52aac60d0da46ce5