Add overloaded beginSession to take in task type
Because starting a new task is equivalent to starting a new session,
successful start of a session indicates that the task successfuly
started. To proceed with hand handling a specified task, we must
know what task type was successfully started.

Thus, we have the task type be a part of the BeginSession command
proto to indicate what type of task should be populated in the newly
created session. This will enable us to read and handle the correct
task on successful response of the BeginSession command.

Bug: 277797528
Test: added
Change-Id: I1de9d5d9619359459e7afa45aa213d1e6066d0e1

GitOrigin-RevId: 3b4b09fcb5673e35cd75ea53340736e4eb219a73