Adjust CpuCaptureSessionArtifact doSelect for task based
The CpuCaptureSessionArtifact doSelect() method will be used for
opening past recordings and imported sessions. Currently, the
doSelect() method early returns if the session is imported. For
the Task Based UX this case needs to be flagged out so that we
can re-use the setAndSelectCapture() call made later in the
function using the sessions start timestamp as the trace id (this
thus enabled us to create cpu capture stages and load imported cpu
sessions).

Bug: 277798530
Test: existing
Change-Id: I9134af106f1526dccb74e884535bf66bd532737f

GitOrigin-RevId: 551a0062bc1ef3cc6171654d986cfeeaf2f0e2e9