Notify wear pairing listeners of current state
Remove the updateDevicesChannel.trySend(Unit) call, which is
insufficient and unnecessary, and instead explicitly notify the listener
of the full pairing state so that it can be fully synchronized with the
WearPairingManager.

While updateDevicesChannel.trySend(Unit) may in some cases send
notifications, it only does so when the PairingState has changed. Adding
a listener is not correlated with any pairing state changes, so there's
no reason to refresh pairing state on account of it.

Bug: 291704385
Test: Added
Change-Id: If201c1917ab836261559ed714e710d07d69fe8ad

GitOrigin-RevId: 75d9a8f3cc93ec219cbf276a821cd0a45c6f148e