Don't allow connecting to a device different than the selected one
Each time an app is launched from Studio, that app's process has the
priority over everything else, and Layout Inspector automatically
connects to it.

This is ok in all cases, except in the embedded Layout Inspector, where
the inspector should connect to a process only if it belongs to the
visible device.

Test: LayoutInspectorProjectServiceTest
Bug: 265150325
Change-Id: I6a9a4888b4386b553d940c5937ef9b8d8f99e77a

GitOrigin-RevId: 90b0144c8cfb757d078ee52127bad656baa6edbf