Refactor Component Tree UI
The component tree UI needs to switch between different states like:
showing the tree, showing the loading indicator, showing messages to
the user, etc.

Currently the switching between different states is done by setting
multiple individual variables. This is fragile and has been source
of problmes over time.

This change refactors the UI so that only one component at a time
can be added to the panel and be visible. Also all individual
variables controlling the UI states are removed, in favor of a
single function responsible for rendering the current UI state.

Bug: 291762878
Bug: 292988420
Test: RootPanelTest
Change-Id: I2ecc63c9e4ce8f6881d88a8eced2b345002d7316

GitOrigin-RevId: 670322c4598ab5c7f309e9826699b3ca97fe5bf3