Update default toolwindow layout provider to new APIs
Besides porting the existing logic for the New UI from the old to the
new APIs, this also adds the Studio Bot toolwindow and hides the App
Quality Insights toolwindow by default, as discussed with csinco@.

The new extension MUST be anchored after the default IJ ones in the
manifest, as we need to first clear what all the default IJ behaviour
(so we can hide things like Maven, Structure, Database, etc by not
adding them back), and then build the layout as we see fit. To achieve
this, the extension is anchored 'last', but could also be anchored as
'after' whatever other extension IJ provides. Given we don't need to
have further extensions though, this is fine for us.

Fixes: 293228537
Change-Id: Ib474739c892190fa7b8b63b3b27bae16a540bd89

GitOrigin-RevId: e9d939ced8d939d229bb0b28f70ef2ea47a853f7