Improve model dumper output for KotlinMPPGradleModel
Platform 232 removed cachedCompilerArgumentsBySourceSet field from
the KotlinMPPGradleModel that represented the compiled arguments using
K2JVMCompilerArguments, end up handled by CommonCompilerArguments
custom dumper that displayed the arguments as properties but also
filtered some of them that weren't considered relevant.

Since we started using the compilerArgumentsBySourceSet the arguments
are represented via List<String>, ignoring the custom dumper. For
this reason a SpecializedDumper was created for this field to be able
to keep same structure we had before on the snapshots.

In addition, The KotlinMPPGradleModel.dependencyMap is not longer
populated requiring to use IdeaKotlinDependenciesContainer.

Bug: 294350954
Test: Existing snapshots
Change-Id: I0338b3ae98feaf1e0e96523a80fc8ce616d43196

GitOrigin-RevId: 15618a27b0ba40116c26529f28e72992e70ebea0