Add TaskStage interface
For reference: the term "task handler" will be used in the
following commit message. This is a concept that will be
introduced in later CLs. In summary, they are classes that
wrap the behavior and some metadata for "task" execution in
the Task-Based UX.

This interface will be implemented in stages that are used to
execute tasks (CpuProfilerStage or MainMemoryProfilerStage).
This now allows the implementing Stage's bound view to invoke
the custom stop action of a task handler.

The reason we need this is because a task handler will own an
instance of the stage, and not the other way around. And so,
the bounded view for the CpuProfilerStage or MainMemoryProfilerStage
(stages that are used for task execution) only has access to the
bound stage, and not the task handler. Thus, as aforementioned,
this now allows us to invoke passed in stop task action behavior
via the stage view without causing a circular reference b/w the
stage and task handler.

Bug: 277798877
Bug: 277798936
Test: added
Change-Id: I9d7c10e656bf4b7ab33541a880e1dabafc14d8a7

GitOrigin-RevId: a3f2e1a1ce29e453bf8f22c29cb8984d6e25b24f