Pass application id to debuggers if known
Native debugging requires the application id of the process that it is
attaching the debugger to. It finds the application id with the help of
`FacetFinder`, which locates the facet and the application id matching
a given `ClientData` object. This process, however, does not work in
Android Studio with Bazel.

When launching an app under the debugger the application id is known
upfront but it is not passed to debuggers. Passing it to debuggers is
sufficient for the native debugger to work in Android Studio with Bazel
when launching applications.

Note: It does not help when attaching to an existing process.

Bug: 292267103
Test: n/a
Change-Id: I4df63de3aceba5bd2e23629d343a32526a244544

GitOrigin-RevId: ab20d576044be581bfd29eabe2b68ece658c6f96