SetAttributeQuickFix: remove write actions
Initiating write actions is unnecessary because apply() is already
called from a write action when the quick-fix is invoked;
this is configured by the return value of startInWriteAction().

And initiating write actions is harmful because it breaks intention
previews, which are generated by calling apply() on a background
thread using a copy of the PSI that can be modified without the
write lock.

Fixes: 295392013
Change-Id: I72ee5296ea4042fd694083475d897370711333ce

GitOrigin-RevId: 2a5e571cdd4dc8a721ade1311a71c92440971fb3