Convert Gradle build root dir to canonical path
When an included build is symlinked, some paths are
canonical and some or not. E.g. build root dir is not,
and project project-dir is. This mismatch causes
issues during sync. This change converts build root dir
to always be canonical.

Note: after fixing this, users will hit IDEA-329756 as
IDEA also has issues when included build is a symlink.
Adding a test is tracked in b/295882834.

Bug: 295337712
Test: manual because of IDEA-329756
Change-Id: I2634cc214b0348c5c2a282294077c44def5e9fcf

GitOrigin-RevId: 7c6d0284104397a17e5125bd15b1a40c2888c747