Add a timeout in all circumstances for gutter drawables
The GutterDrawableFactory is used by the autocomplete to display a small
icon next to a drawable id. The icon generation may require a read
action, but the autocomplete displays while typing which requires a
write action. That can cause a deadlock. To avoid this, this adds a
timeout to the icon rendering.

Bug: 295049594
Test: N/A
Change-Id: I5e8414fa91e872458fb03ef798ab2de9f8b3ba9a

GitOrigin-RevId: cf1c98125e848c08e9b023a9af3dd073f0455236