Add insertion of CPU_TRACE event at import-time
Currently, a CPU_TRACE event (backing event to generate a
CpuCaptureSessionArtifact within an session) is inserted
when the the Profiler detects a session type change to the
CPU_CAPTURE type (which is triggered when a CPU file is
imported). However, in the Task-Based UX, we need to insert
the respective CPU_TRACE event and thus consequent artifact
at import time, rather than wait for the CpuCaptureStage to
be entered.

Thus, this change introduces CPU_TRACE event insertion at
import-time and thus also disables the insertion done on
entering CpuCaptureStage. We can disable the CPU_CAPTURE
session change listener with confidence as well as the
Task-Based UX does not enter any stage or perform any
responsive action on import.

Bug: 277798530
Test: added + updated + manual
Change-Id: Iceb41ae5fa6e09f53b34cd3a255ae4879cacec04

GitOrigin-RevId: c246afd291184b098268f82a1f166c14379c5591