Rename and refactor initializeTaskTabConent method
...to better reflect that its initializing the entire tab (tab
itself + content). Also, usages have been moved so that it is
called on initialization of the tool window along with the home
tab. As documented in the code, the Profiler tab is initialized
here with the home tab so that the view bindings will be ready
in the case the user imports a file from a fresh/un-opened
Profiler tool window state. While entering a stage from an
uninitialized Profiler state after importing is not a possible
flow in the Task-Based UX, the initialization of the Profiler tab
logic is used for both the Sessions-based Profiler tab and the
Task-Based UX Profiler tab, so it must be called in a place that
accommodates both tabs.

Bug: 277797528
Test: existing
Change-Id: I4fc8ef1e66508057ed2f920b41dd62b34f2d58e1

GitOrigin-RevId: 4e109baf2da5dc30ee06aa708ca2ccd711d7be78