Add Garbage collection
.. in toolbar and context menu of memory.

Message displayed in toolbar is moved below toolbar and garbage icon
is made to display on the leftmost of toolbar

Garbage collection code from BaseMemoryTimelineComponent and
MainMemoryProfilerStageView is moved to GarbageCollectionComponent(new
class), so code can also be accessed by LiveMemoryAllocationView.
Similarly, gc code from BaseStreamingMemoryProfilerStage is moved to
MemoryDataProvider.

Bug: 277799031
Test: Added
Change-Id: Ia7325ceb96e08fb48f13cfd819b8637f58a86e6d

GitOrigin-RevId: 4e843dcb1b54730f8ee6a42a496a84b8dd7fe810