Add project descriptors to test fixtures.
These descriptors encapsulate the Java language level, JDK, and required
libraries that a project fixture needs to add to all of its created
modules. The required libraries are especially important for Kotlin
code, which needs kotlin-stdlib on the classpath to work correctly (or,
thanks to a bug in kotlinc 1.9.20-dev, to even be _analyzed_.)

The `AdtTestProjectDescriptors` object contains a stock descriptor for
Java-only projects that sets the language level and JDK, and two for
Kotlin projects that add kotlin-stdlib (one with accompanying srcjars,
and one without). It also defines a default descriptor that will force
kotlin-stdlib to load when the K2 plugin is in use, to work around the
aforementioned kotlinc bug.

The descriptor logic, and the artifact-finding logic in
`AdtTestKotlinArtifacts`, supersede the equivalent code in upstream
IntelliJ kotlin.test-framework. The ADT versions of these classes have
adaptations to allow them to work with the prebuilt layout used in
Android Studio tests, rather than the unpacked source layout used in
upstream IJ tests.

This reverts commit da29c4f5307c5362d99ea2c34837cc2547192ffa, now that
the tests that were disabled by that commit are fixed.

Fixes: 294248298
Test: bazel test --test_tag_filters=kotlin-plugin-k2 //tools/adt/idea/...
Change-Id: I5edac1decdb72c8a55f687c4d153bc7f877775fe

GitOrigin-RevId: 83790c71bc0a09b12001109a24b3c029d8397b1c