[android] IDEA-330091 Introducing a Feature Flag to Ease Android Gradle Plugin Compatibility Restrictions
In IntelliJ IDEA, we currently have a problem that the Android Gradle Plugin (AGP)
version supported by IDEA is always one version behind Android Studio.
At the time of this commit, the latest supported AGP version by IDEA was 8.1.0.

The trailing behind the Android Studio AGP version causes problems for users
who want to use the latest AGP versions in their Android and KMM apps.
Because of this version trailing problem,
they cannot develop Android / KMM apps with the latest available AGP in IntelliJ IDEA.

With this change, we want to introduce a feature flag that, when enabled,
relaxes the AGP compatibility check in the Android plugin in the following way:
Compatibility between the IDE and project AGP versions will be maintained
if the following conditions are met:
- The major AGP version of the project corresponds to the major AGP version of the IDE.
- The minor AGP version of the IDE is less than or equal to the minor AGP version of the project, incremented by one.
- Version distinctions such as snapshot, alpha, and beta designations will be disregarded in this context.

This feature flag may cause build failures due to API incompatibility
between the current project and the existing IDE-supported AGP version.
That's why we consider it experimental and internal,
and we'll revisit it one more time closer to the 233 release.

GitOrigin-RevId: 02240808f898f052bff12159e798c307b29e986e