Remove call to auto-select added tab
To prevent invocation of every tab's select handler
every time its recreated and added. The recreation
and addition of tabs happens any time there is a
change in tab selection so this can become costly,
especially if the on select handler performs heavy
computation. Worth noting that this call to set
the selection when adding the tab is a side-effect
and unexpected behavior.

The select handler is still registered when adding the
tab, but now requires an explicit selection to be
triggered which is expected behavior.

One test was removed (tabCallbackIsCalledOnSelected)
because it was not actually triggering a select event.
It was succeeding because of the explicit call to
select the tab when it was being added. I tried to
invoke the selection via ui in different way (via
invoking its explicit mouse listener, FakeUI mouse click,
etc. to no avail).

Bug: 281850040
Test: added + updated
Change-Id: Id7cc91fd14c3b57596cda8540530dd26afc32d8d

GitOrigin-RevId: 43eb4cc30c74f36d4e64aca475f60f4b992f211a