Fix SimpleHighlightingBenchmark
This perfgate test was failing. The underlying cause is that the
function InspectionProfileImpl.getAllEnabledInspectionTools was called
*before* loading the test project.

Details:

This initializes the inspection profile for the project, but at that
moment the project is an artificial one generated for the unit test by
AndroidGradleProjectRule. So the inspection profile is initialized
against that project, and once the real project is loaded and
highlighting is requested for a file in it, there are all sorts of
misconfigurations.

The error that ended up causing the test to crash was because the
available ProjectSystem was a DefaultProjectSystem rather than a
GradleProjectSystem, so the manifest was not able to be parsed and
resource references didn't resolve.

Bug: 297025863
Test: manual, the test no longer fails
Change-Id: Ie6d854eb7cfefe0a7d5aebbd3397fa5c287aaacd

GitOrigin-RevId: 4704f45cbd5a85d00c600f5d0f510a279414dffa