IDEA-160570 Pass project to SdkSync to determine if it's an Android project
Updated the SdkSync class to receive a project parameter and determine if the project is an Android project using the presence of the AndroidManifest.xml file.
This change helps IDEA to avoid creating local.properties in non-Android projects. Modified related methods and classes to pass the project as an argument to SdkSync.

GitOrigin-RevId: 7e6141689cc8ff1aed02181f4f3aa03809290ade