Remove redraw on capture node hover
The dataUpdated flag in the HTreeChart would trigger
a redraw of the entire chart if set to true. One of
the places it was set to true is when the focused or
currently moused over capture node changed. This
resulted in every mouse movment from node-to-node
causing a ~200-250 ms long redraw if only a
call to render was made, or ~300 ms if nodes and
rectangles needed to be reconstructed in addition
to the call to render.

The only value of this redraw for a node focus change
is that the hovered over node has a slightly tinted
fill color. However, after discussing it with UX and
product side, it was decided that we would remove the
hover fill color for the performance enhacement. Thus,
dataUpdated and the consequent redraw triggered has
been removed.

One caveat of removing the trigger to redraw on mouse
hover is that, if something else (like a timeline
range change) does trigger a redraw, we do not want to
show a different fill color on the last hovered node.
Thus, theisFocused parameter of the capture node render
method is now statically set as false to prevent all
hover coloring. This achieves a consistent UI (if a
mouse position change does not update the node's fill
color, no other chart update should either).

Fixes: 281850040
Test: manual
Change-Id: If579cb5f862d62806d2ba9da4c17343ebf3087c7

GitOrigin-RevId: b6bccb9579d13d04b3d3f40ed93190fdebfa844a