Convert ArtifactCoordinate to an enum
In production, ArtifactCoordinate is used for only two (and a half)
artifacts: androidx's work-runtime and Compose ui (and the post-KMP
ui-android).  Those artifacts provide a minimum version for querying a
running process, though results from this query are largely ignored
for Android Studio with Blaze.  For non-ASwB, the results of this
query are used for resolving the artifact.

Make the different semantics explicit by providing: an enum class for
the specific artifacts with their minimum versions; conversion
functions for the occasions where we need a wild version ("+") for
ASwB to be able to connect to arbitrary versions, or an any-version
specifier for querying what the available library version is; and a
data class to represent the coordinate resulting from querying the
running application.

Preserve the appearance of an ArtifactCoordinate() constructor for use
in tests by defining an invoke() operator function in the interface's
companion object.  (Brought to you by IDE-suggestion-oriented
development).

There are a few tests which are slightly weakened in this change; the
stronger variant of those tests is restored in the next change.

Bug: 279886738
Test: existing tests adjusted
Change-Id: I2d79e86e2c375bd42f25985f0d3dde3e1a1b7da2

GitOrigin-RevId: 2d57d9258cba18b7aa99274ba3ba311943062679